1087345252736090730
ManifestID: 1087345252736090730
Date: 5 October 2023 – 17:53:17 UTC
Version: U49-575720

+++ Assembly-CSharp/Klei/AI/Disease.cs
@@ -19,7 +19,7 @@ namespace Klei.AI
 			}
 		}
 
-		public Disease(string id, byte strength, Disease.RangeInfo temperature_range, Disease.RangeInfo temperature_half_lives, Disease.RangeInfo pressure_range, Disease.RangeInfo pressure_half_lives, float radiation_kill_rate, bool statsOnly)
+		public Disease(string id, float strength, Disease.RangeInfo temperature_range, Disease.RangeInfo temperature_half_lives, Disease.RangeInfo pressure_range, Disease.RangeInfo pressure_half_lives, float radiation_kill_rate, bool statsOnly)
 			: base(id, null, null)
 		{
 			this.name = new StringKey("STRINGS.DUPLICANTS.DISEASES." + id.ToUpper() + ".NAME");
